Case 27689; removed duplication of options
