+++ src/Command/Site/Checkout/CheckoutCommand.php
@@ -10,25 +10,15 @@
 namespace DennisDigital\Drupal\Console\Command\Site\Checkout;
 
 use Symfony\Component\Console\Input\ArrayInput;
-use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
-use Symfony\Component\Console\Exception\InvalidOptionException;
-use DennisDigital\Drupal\Console\Command\Site\AbstractCommand;
 
 /**
  * Class SiteCheckoutCommand
  *
  * @package DennisDigital\Drupal\Console\Command
  */
-class CheckoutCommand extends AbstractCommand {
-  /**
-   * Types of refs that can be checked out.
-   *
-   * @var array
-   */
-  protected $refTypes = ['tag', 'branch'];
-
+class CheckoutCommand extends AbstractCheckoutCommand {
   /**
    * {@inheritdoc}
    */
@@ -37,24 +27,6 @@ class CheckoutCommand extends AbstractCommand {
 
     $this->setName('site:checkout')
       ->setDescription('Checkout a repository');
-
-    // Custom options.
-    $this->addOption(
-      'force',
-      '',
-      InputOption::VALUE_NONE,
-      'Will force the checkout and replace all local changes'
-    );
-
-    // Allow different ref types to be checked out.
-    foreach ($this->refTypes as $refType) {
-      $this->addOption(
-        $refType,
-        '-' . strtoupper(substr($refType, 0, 1)),
-        InputOption::VALUE_OPTIONAL,
-        sprintf('Specify which %s to checkout.', $refType)
-      );
-    }
   }
 
   /**
@@ -62,16 +34,13 @@ class CheckoutCommand extends AbstractCommand {
    */
   protected function execute(InputInterface $input, OutputInterface $output) {
     // Get command based one branch/tag/revision option.
+    $command = $this->getApplication()->find('site:checkout:branch');
     foreach ($this->refTypes as $refType) {
       if ($input->hasOption($refType) && $input->getOption($refType) !== NULL) {
         $commandName = 'site:checkout:' . $refType;
         $command = $this->getApplication()->find($commandName);
         $input->setArgument('command', $commandName);
       }
-    }
-
-    if (!isset($command)) {
-      $command = $this->getApplication()->find('site:checkout:branch');
     }
 
     // Pass input parameters to specific checkout command.
